Authorize ations using User's model

+++ app/Http/Controllers/Api/V2/TaskController.php
@@ -3,10 +3,11 @@
 namespace App\Http\Controllers\Api\V2;
 
 use App\Models\Task;
+use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Http\Resources\TaskResource;
 use App\Http\Requests\StoreTaskRequest;
 use App\Http\Requests\UpdateTaskRequest;
-use App\Http\Resources\TaskResource;
 
 class TaskController extends Controller
 {
@@ -15,6 +16,10 @@ class TaskController extends Controller
      */
     public function index()
     {
+        if ($request->user()->cannot('viewAny', Task::class)) {
+            abort(403);
+        }
+
         // return TaskResource::collection(Task::all());
         return request()->user()
             ->tasks()
@@ -27,7 +32,10 @@ class TaskController extends Controller
      */
     public function store(StoreTaskRequest $request)
     {
-        // $task = Task::create($request->validated() + ['user_id' => $request->user()->id]);
+        if ($request->user()->cannot('create', Task::class)) {
+            abort(403);
+        }
+        
         $task = $request->user()->tasks()->create($request->validated());
 
         return $task->toResource();
@@ -36,10 +44,12 @@ class TaskController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Task $task)
+    public function show(Task $task, Request $request)
     {
-        // return new TaskResource($task);
-        // return TaskResource::make($task);
+        if ($request->user()->cannot('view', $task)) {
+            abort(403);
+        }
+
         return $task->toResource();
     }
 
@@ -48,6 +58,10 @@ class TaskController extends Controller
      */
     public function update(UpdateTaskRequest $request, Task $task)
     {
+        if ($request->user()->cannot('update', $task)) {
+            abort(403);
+        }
+
         $task->update($request->validated());
 
         return $task->toResource();
@@ -56,8 +70,12 @@ class TaskController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task, Request $request)
     {
+        if ($request->user()->cannot('delete', $task)) {
+            abort(403);
+        }
+
         $task->delete();
 
         return response()->noContent();
